[grsec] Applied patch 3.1-201606051644
This is grsecurity-3.1-4.5.6-201606051644.patch

+++ drivers/pci/probe.c
@@ -182,7 +182,7 @@ int __pci_read_base(struct pci_dev *dev, enum pci_bar_type type,
 	if (dev->non_compliant_bars)
 		return 0;
 
-	mask = type ? PCI_ROM_ADDRESS_MASK : ~0;
+	mask = type ? (u32)PCI_ROM_ADDRESS_MASK : ~0;
 
 	/* No printks while decoding is disabled! */
 	if (!dev->mmio_always_on) {
